#319
Autocompleting with typing just 2 letters of toppings or else If any users want to add a new topping then just type and it will be syncing with toppings. At the time of editing the food, it will automatically sync with the topping taxonomy from the front end and back end.